refactor: move resolution-related files under `scales/`
