+++ packages/core/src/gl/webGLHelper.js
@@ -231,7 +231,7 @@ export default class WebGLHelper {
      *
      * TODO: This may be too specific to be included in WebGLHelper. Find a better place.
      *
-     * @param {import("../view/scaleResolution.js").default} resolution
+     * @param {import("../scales/scaleResolution.js").default} resolution
      * @param {boolean} update Update the texture if it exists already.
      */
     createRangeTexture(resolution, update = false) {
