+++ packages/core/src/scales/scaleDomainAggregator.js
