+++ packages/core/src/scales/scaleInteractionController.js
