+++ packages/core/src/scales/scalePropsResolver.js
