+++ packages/core/src/scales/scaleResolutionConstants.js
