+++ packages/core/src/scales/scaleRules.js
